refactor: msw package 삭제/mock 삭제/serviceworker 삭제
