+++ package.json
@@ -42,7 +42,6 @@
     "eslint-plugin-react-refresh": "^0.4.3",
     "husky": "^8.0.0",
     "lint-staged": "^15.0.2",
-    "msw": "^2.0.0",
     "prettier": "^3.0.3",
     "typescript": "^5.0.2",
     "vite": "^4.4.5"
